mcp_servers/powershell_server/requirements.txt
mcp>=1.0.0
pywinrm>=0.4.3
requests-ntlm>=1.2.0